Consolidated reaction propagator files
Function-based and class-based all on the same file now.
